モデル整合性チェック用に改造

+++ raspi/main_project/datasets/train_model.py
@@ -4,11 +4,16 @@ import logging
 import os
 
 logger = logging.getLogger(__name__)
-script_dir = os.path.dirname(os.path.abspath(__file__))
-model_path = os.path.join(script_dir, "ft_model.bin")
 
 def train_model():
-    df = pd.read_csv(f"{script_dir}/gakusyudata.csv")
+    logger.info("分類モデルの確認開始...")
+    script_dir = os.path.dirname(os.path.abspath(__file__))
+    model_path = os.path.join(script_dir, "ft_model.bin")
+    csv_path = os.path.join(script_dir, "gakushudata.csv")
+    if os.path.exists(model_path):
+        logger.info("分類モデルが存在します")
+        return
+    df = pd.read_csv(csv_path)
 
     with open("train.txt","w",encoding="utf-8") as f:
         for _, row in df.iterrows():
